requirements.txt
numpy
pandas
matplotlib
seaborn
psycopg2
python-dotenv
-e .